backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "API para e-commerce de joias"
authors = [
    {name = "Luiz Fernando da Silva Guedes",email = "pitohuikun@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=6.0,<7.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "django-filter (>=25.2,<26.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
